Feature/bootstrap configuration konvoydp

+++ components/konvoy-control-plane/pkg/config/loader_test.go
@@ -73,9 +73,9 @@ apiServer:
 		Expect(cfg.XdsServer.DiagnosticsPort).To(Equal(5003))
 
 		Expect(cfg.BootstrapServer.Port).To(Equal(5004))
-		Expect(cfg.BootstrapServer.Params.AdminPort).To(Equal(1234))
+		Expect(cfg.BootstrapServer.Params.AdminPort).To(Equal(uint32(1234)))
 		Expect(cfg.BootstrapServer.Params.XdsHost).To(Equal("konvoy-control-plane"))
-		Expect(cfg.BootstrapServer.Params.XdsPort).To(Equal(4321))
+		Expect(cfg.BootstrapServer.Params.XdsPort).To(Equal(uint32(4321)))
 
 		Expect(cfg.Environment).To(Equal(konvoy_cp.KubernetesEnvironment))
 
@@ -130,9 +130,9 @@ apiServer:
 		Expect(cfg.XdsServer.DiagnosticsPort).To(Equal(5003))
 
 		Expect(cfg.BootstrapServer.Port).To(Equal(5004))
-		Expect(cfg.BootstrapServer.Params.AdminPort).To(Equal(1234))
+		Expect(cfg.BootstrapServer.Params.AdminPort).To(Equal(uint32(1234)))
 		Expect(cfg.BootstrapServer.Params.XdsHost).To(Equal("konvoy-control-plane"))
-		Expect(cfg.BootstrapServer.Params.XdsPort).To(Equal(4321))
+		Expect(cfg.BootstrapServer.Params.XdsPort).To(Equal(uint32(4321)))
 
 		Expect(cfg.Environment).To(Equal(konvoy_cp.KubernetesEnvironment))
